move and rename PySM operator to map
